Done importing the data from the API

+++ lib/movie/movie.dart
@@ -19,15 +19,15 @@ Movie({
 });
 
 
-factory Movie.fromJson(Map<String, dynamic> json){
-  return Movie(
-   title: json["title"],
-   backDropPath: json["backdrop_path"],
-   originalTitle: json["original_title"],
-   overview: json["overview"],
-   posterPath: json["poster_path"],
-   releaseDate: json["release_date"],
-   voteAverage: json["vote_average"]
-   );
+  factory Movie.fromJson(Map<String, dynamic> json) {
+    return Movie(
+      title: json["title"],
+      backDropPath: json["backdrop_path"],
+      originalTitle: json["original_title"],
+      overview: json["overview"],
+      posterPath: json["poster_path"],
+      releaseDate: json["release_date"],
+      voteAverage: json["vote_average"].toDouble(),
+    );
 }
 }
